Update fcpy.sh
Remove the "exit 0" so the script continues and the container works right off the bat. Instead of kill the container and restarting a stopped container. 
With this change this will also work in a docker stack as a service. Since a service will never restart a stopped container but spin up a new one.